+++ processor/src/test/fixtures/input/com/example/route/RouteWithEmptyPath.java
@@ -1,0 +1,10 @@
+package com.example.route;
+
+import router.fu.annotations.Route;
+import router.fu.annotations.Router;
+
+@Router
+@Route( name = "welcome", path = "" )
+public class RouteWithEmptyPath
+{
+}