+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Upgrade the `com.squareup:javapoet` artifact to version `1.13.0`.
+* Update the annotation processor so that the hash suffixes like `"#/"`, `"#"` and `""` are treated as an empty path.
 
 ### [v0.29](https://github.com/realityforge/router-fu/tree/v0.29) (2020-06-18) · [Full Changelog](https://github.com/realityforge/router-fu/compare/v0.28...v0.29)